requirements.txt
# Core
GitPython==3.1.42
pyyaml==6.0.1

# Linting helpers (optional rules may use these)
dockerfile-parse==2.0.0
yamllint==1.33.0

# CLI & Dev
rich==13.7.1        # for better CLI output in future
termcolor==2.4.0    # optional: for colorful prints

